+++ Chapter.Net.WPF.Theming/Enums/WindowTheme.cs
@@ -1,0 +1,31 @@
+﻿// -----------------------------------------------------------------------------------------------------------------
+// <copyright file="WindowTheme.cs" company="my-libraries">
+//     Copyright (c) David Wendland. All rights reserved.
+// </copyright>
+// -----------------------------------------------------------------------------------------------------------------
+
+// ReSharper disable once CheckNamespace
+
+namespace Chapter.Net.WPF.Theming
+{
+    /// <summary>
+    ///     The theme to use.
+    /// </summary>
+    public enum WindowTheme
+    {
+        /// <summary>
+        ///     Light mode.
+        /// </summary>
+        Light,
+
+        /// <summary>
+        ///     Dark mode.
+        /// </summary>
+        Dark,
+
+        /// <summary>
+        ///     Same as system.
+        /// </summary>
+        System
+    }
+}