+++ Demo/App.xaml.cs
@@ -1,0 +1,11 @@
+﻿// -----------------------------------------------------------------------------------------------------------------
+// <copyright file="App.xaml.cs" company="my-libraries">
+//     Copyright (c) David Wendland. All rights reserved.
+// </copyright>
+// -----------------------------------------------------------------------------------------------------------------
+
+namespace Demo;
+
+public partial class App
+{
+}