+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added theme manager to get and set the window theme either explicit or by system.
 ### Supported .Net Versions
 - .Net Core 3.0
 - .Net Framework 4.5